renamed folder, configured csbc project
