+++ portal-configurations/csbc-pson/docTitleConfig.ts
@@ -1,7 +1,7 @@
 import { DocTitleConfig } from '../types/portal-config'
 
 const docTitleConfig: DocTitleConfig = {
-  name: 'CSBC-PS/ON'
+  name: 'CSBC  /  PS-ON'
 }
 
 export default docTitleConfig
