+++ portal-configurations/csbc-pson/homeExploreConfiguration/data.ts
@@ -1,5 +1,6 @@
 import { SynapseConstants } from 'synapse-react-client'
 import { HomeExploreConfig } from '../../types/portal-config'
+import loadingScreen from '../loadingScreen'
 
 // tslint:disable-next-line:max-line-length
 const sql = 'SELECT grantType AS "Grant Type", centerName AS "Grant", consortium AS "Program", species AS "Species", fileFormat AS "Data Format", experimentalStrategy AS "Assay", platform AS "Platform", tumorType AS "Disease Type", sex AS "Gender", tissue AS "Tissue", name as "File Name"  FROM syn9630847'
@@ -30,6 +31,7 @@ export const data: HomeExploreConfig = {
       rgbIndex,
       facetName,
       unitDescription,
+      loadingScreen,
       name: 'Data',
       facetAliases: {
         grantType: 'Grant Type',
@@ -51,6 +53,7 @@ export const data: HomeExploreConfig = {
     name: 'QueryWrapperMenu',
     props: {
       rgbIndex,
+      loadingScreen,
       menuConfig: [
         {
           sql,
