+++ portal-configurations/csbc-pson/loadingScreen.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react'
 import { BarLoader } from 'react-spinners'
 
-const loadingBar = <div className="bar-loader"><BarLoader color="#47337D" loading={true}/></div>
+const loadingScreen = <div className="bar-loader"><BarLoader color="#47337D" loading={true}/></div>
 
-export default loadingBar
+export default loadingScreen
